rearrange preparser
